Fix 2 warnings in python/python-membuf.c
2009-02-24 Phil Muldoon <pmuldoon@redhat.com>

	* python/python-membuf.c (get_char_buffer): Initialize buffer to NULL.
	* python/python.c (add_pattern_element): Use Py_ssize_t.

+++ gdb/python/python.c
@@ -387,7 +387,7 @@ add_pattern_element (PyObject *obj, int size, char **pattern_buf,
          object implementing the buffer protocol.  The SIZE parameter is
 	 ignored in this case.  */
 
-      int val_bytes;
+      Py_ssize_t val_bytes;
       const void *buffer;
 
       if (PyObject_AsReadBuffer (obj, &buffer, &val_bytes) == -1)
